# pkg/server/server.go
package server

import (
	"2103_proto_f_server/pkg/db"
	"2103_proto_f_server/pkg/http/middleware"
	"2103_proto_f_server/pkg/http/response"
	"2103_proto_f_server/pkg/server/handler"
	"2103_proto_f_server/pkg/server/model"
	"2103_proto_f_server/pkg/server/service"
	"fmt"
	"log"
	"net/http"
)

var (
	httpResponse = response.NewHttpResponse()

	userRepository = model.NewUserRepository(db.Conn)
	authMiddleware = middleware.NewMiddleware(httpResponse, userRepository)

	userService    = service.NewUserService(userRepository)
	gameService    = service.NewGameService(userRepository)
	rankingService = service.NewRankingService(userRepository)

	userHandler    = handler.NewUserHandler(httpResponse, userService)
	gameHandler    = handler.NewGameHandler(httpResponse, gameService)
	rankingHandler = handler.NewRankingHandler(httpResponse, rankingService)
)

// Serve HTTPサーバを起動する
func Serve(addr string) {
	http.HandleFunc("/hello", func(w http.ResponseWriter, r *http.Request) {
		fmt.Fprintf(w, "Hello World")
	})

	http.HandleFunc("/user/create", post(userHandler.HandleUserCreate))
	http.HandleFunc("/user/get", get(authMiddleware.Authenticate(userHandler.HandleUserGet)))
	http.HandleFunc("/game/finish", post(authMiddleware.Authenticate(gameHandler.HandleGameFinish)))
	http.HandleFunc("/ranking/list", get(rankingHandler.HandleRankingList))

	/* ===== サーバの起動 ===== */
	log.Println("Server running...")
	err := http.ListenAndServe(addr, nil)
	if err != nil {
		log.Fatalf("Listen and serve failed. %+v", err)
	}
}

// get GETリクエストを処理する
func get(apiFunc http.HandlerFunc) http.HandlerFunc {
	return httpMethod(apiFunc, http.MethodGet)
}

// post POSTリクエストを処理する
func post(apiFunc http.HandlerFunc) http.HandlerFunc {
	return httpMethod(apiFunc, http.MethodPost)
}

// httpMethod 指定したHTTPメソッドでAPIの処理を実行する
func httpMethod(apiFunc http.HandlerFunc, method string) http.HandlerFunc {
	return func(writer http.ResponseWriter, request *http.Request) {

		// CORS対応
		writer.Header().Add("Access-Control-Allow-Origin", "*")
		writer.Header().Add("Access-Control-Allow-Headers", "Content-Type,Accept,Origin,x-token")

		// プリフライトリクエストは処理を通さない
		if request.Method == http.MethodOptions {
			return
		}
		// 指定のHTTPメソッドでない場合はエラー
		if request.Method != method {
			writer.WriteHeader(http.StatusMethodNotAllowed)
			writer.Write([]byte("Method Not Allowed"))
			return
		}

		// 共通のレスポンスヘッダを設定
		writer.Header().Add("Content-Type", "application/json")
		apiFunc(writer, request)
	}
}